refactor: 앱과 페이지 코드 분리
